+++ types/index.d.ts
@@ -1,3 +1,4 @@
+export function clamp(val: number, min: number, max: number): number;
 export function round(number: number, precision?: number): number;
 export class Email extends URL {
 	/**